refactor(settings): move AppSettings + Settings screens to impl, delegate via BlocScreen
AppSettingsScreen, AppSettingsPreviews, and SettingsScreen move from
settings/public to settings/impl/ui. SettingsScreen's inline
preview Blocs (3 variants — unauthenticated / authenticated /
anonymous) gain `Content()` overrides via replace_all.

AppSettingsBloc and SettingsBloc both extend BlocScreen; the
respective Impls override Content() to call their screens.

RootBloc.Child.AppSettings and BottomNavBloc.Child.Settings gain
`BlocScreen by bloc`; the corresponding `when` branches in
RootScreen and BottomNavScreen render via child.Content() /
instance.Content().

`SettingsRow` stays internal in SettingsScreen — both AppSettings
and Settings live in the same impl/ui package now, so the same-
module visibility keeps working without exposing it publicly.

settings/public sets publicResClass = true; settings/impl gains
the Compose plugin + ui/text/compose-resources deps.
