+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavBloc.kt
@@ -13,6 +13,7 @@ import com.plusmobileapps.chefmate.recipe.core.addmeal.MealPlannerRootBloc
 import com.plusmobileapps.chefmate.recipe.data.ExtractedRecipeData
 import com.plusmobileapps.chefmate.recipe.list.RecipeListBloc
 import com.plusmobileapps.chefmate.settings.SettingsBloc
+import com.plusmobileapps.chefmate.ui.BlocScreen
 import kotlinx.coroutines.flow.StateFlow
 
 interface BottomNavBloc : BackHandlerOwner, BackClickBloc {
@@ -43,7 +44,7 @@ interface BottomNavBloc : BackHandlerOwner, BackClickBloc {
 
         data class Browser(val bloc: BrowserRootBloc) : Child()
 
-        data class Settings(val bloc: SettingsBloc) : Child()
+        data class Settings(val bloc: SettingsBloc) : Child(), BlocScreen by bloc
     }
 
     sealed class Output {
