+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavScreen.kt
@@ -50,7 +50,6 @@ import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.MEALS
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.RECIPES
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.SETTINGS
 import com.plusmobileapps.chefmate.recipe.list.RecipeListScreen
-import com.plusmobileapps.chefmate.settings.SettingsScreen
 import com.plusmobileapps.chefmate.text.asTextData
 import com.plusmobileapps.chefmate.ui.components.NavRailItem
 import com.plusmobileapps.chefmate.ui.components.PlusNavRailHeaderContainer
@@ -157,7 +156,7 @@ private fun BottomNavContentContainer(bloc: BottomNavBloc, modifier: Modifier = 
             is BottomNavBloc.Child.GroceryList -> GroceryListScreen(instance.bloc)
             is BottomNavBloc.Child.Meals -> MealPlanScreen(instance.bloc)
             is BottomNavBloc.Child.RecipeList -> RecipeListScreen(instance.bloc)
-            is BottomNavBloc.Child.Settings -> SettingsScreen(instance.bloc)
+            is BottomNavBloc.Child.Settings -> instance.Content()
         }
     }
 }
