+++ client/settings/public/build.gradle.kts
@@ -16,4 +16,6 @@ kotlin {
     }
 }
 
+compose { resources { publicResClass = true } }
+
 plusLibrary { namespace = "com.plusmobileapps.chefmate.settings" }
