+++ client/settings/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/settings/SettingsBloc.kt
@@ -3,9 +3,10 @@ package com.plusmobileapps.chefmate.settings
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.text.TextData
+import com.plusmobileapps.chefmate.ui.BlocScreen
 import kotlinx.coroutines.flow.StateFlow
 
-interface SettingsBloc {
+interface SettingsBloc : BlocScreen {
     val state: StateFlow<Model>
 
     fun onSignInClicked()
