+++ client/ui/screenshot-test/build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
     implementation(project(":client:recipe:list:public"))
     implementation(project(":client:recipe:data:public"))
     implementation(project(":client:settings:public"))
+    implementation(project(":client:settings:impl"))
 
     screenshotTestImplementation(libs.screenshot.validation.api)
     screenshotTestImplementation(libs.androidx.compose.ui.tooling)
