+++ client/ui/screenshot-test/src/screenshotTest/kotlin/com/plusmobileapps/chefmate/ui/screenshot/AppSettingsScreenshotTest.kt
@@ -4,8 +4,7 @@ import android.content.res.Configuration
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
-import com.plusmobileapps.chefmate.settings.AppSettingsScreen
-import com.plusmobileapps.chefmate.settings.previewAppSettingsBloc
+import com.plusmobileapps.chefmate.settings.impl.ui.previewAppSettingsBloc
 import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
 
 // Regression coverage for PR #161: the App Settings screen used PlusNavContainer (no themed
@@ -16,12 +15,12 @@ import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
 @Preview(showBackground = true)
 @Composable
 fun AppSettingsLightScreenshot() {
-    ChefMateTheme { AppSettingsScreen(bloc = previewAppSettingsBloc) }
+    ChefMateTheme { previewAppSettingsBloc.Content() }
 }
 
 @PreviewTest
 @Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
 fun AppSettingsDarkScreenshot() {
-    ChefMateTheme(darkTheme = true) { AppSettingsScreen(bloc = previewAppSettingsBloc) }
+    ChefMateTheme(darkTheme = true) { previewAppSettingsBloc.Content() }
 }
